+++ src/main/java/io/github/astrapi69/spring/service/api/GenericService.java
@@ -24,7 +24,10 @@
  */
 package io.github.astrapi69.spring.service.api;
 
+import java.io.Serializable;
 import java.util.Optional;
+
+import javax.persistence.EntityManager;
 
 import lombok.NonNull;
 
@@ -33,7 +36,9 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GenericService<ENTITY, ID, REPOSITORY extends JpaRepository<ENTITY, ID>>
+import io.github.astrapi69.entity.identifiable.Identifiable;
+
+public interface GenericService<ENTITY extends Identifiable<ID>, ID extends Serializable, REPOSITORY extends JpaRepository<ENTITY, ID>>
 {
 	default long count()
 	{
@@ -70,10 +75,16 @@ public interface GenericService<ENTITY, ID, REPOSITORY extends JpaRepository<ENT
 		getRepository().deleteAllInBatch(entities);
 	}
 
+	default boolean exists(@NonNull ENTITY entity)
+	{
+		return existsById(entity.getId());
+	}
+
 	default boolean existsById(@NonNull ID id)
 	{
 		return getRepository().existsById(id);
 	}
+
 
 	default Iterable<ENTITY> findAll()
 	{
@@ -105,6 +116,26 @@ public interface GenericService<ENTITY, ID, REPOSITORY extends JpaRepository<ENT
 		getRepository().flush();
 	}
 
+	default ENTITY getById(@NonNull ID id)
+	{
+		return getRepository().getById(id);
+	}
+
+	/**
+	 * Returns a reference to the entity with the given identifier. Depending on how the JPA
+	 * persistence provider is implemented this is very likely to always return an instance and
+	 * throw an {@link javax.persistence.EntityNotFoundException} on first access. Some of them will
+	 * reject invalid identifiers immediately.
+	 *
+	 * Note: will be removed on next minor release.
+	 *
+	 * @param id
+	 *            must not be {@literal null}.
+	 * @return a reference to the entity with the given identifier.
+	 * @see EntityManager#getReference(Class, Object) for details on when an exception is thrown.
+	 * @deprecated use {@link GenericService#getById(ID)} instead.
+	 */
+	@Deprecated
 	default ENTITY getOne(@NonNull ID id)
 	{
 		return getRepository().getById(id);
@@ -125,6 +156,11 @@ public interface GenericService<ENTITY, ID, REPOSITORY extends JpaRepository<ENT
 	default ENTITY saveAndFlush(@NonNull ENTITY entity)
 	{
 		return getRepository().saveAndFlush(entity);
+	}
+
+	default ENTITY update(@NonNull ENTITY entity)
+	{
+		return save(entity);
 	}
 
 }